refactor into two components
