--- src/pages/[library]/login/[[...methodId]].tsx
@@ -1,24 +1,0 @@
-import * as React from "react";
-import { NextPage, GetStaticPaths, GetStaticProps } from "next";
-import LayoutPage from "components/LayoutPage";
-import withAppProps, { AppProps } from "dataflow/withAppProps";
-import Login from "auth/Login";
-
-const LoginPage: NextPage<AppProps> = ({ library, error }) => {
-  return (
-    <LayoutPage library={library} error={error}>
-      <Login />
-    </LayoutPage>
-  );
-};
-
-export const getStaticProps: GetStaticProps = withAppProps();
-
-export const getStaticPaths: GetStaticPaths = async () => {
-  return {
-    paths: [],
-    fallback: true
-  };
-};
-
-export default LoginPage;
